+++ src/formatting/global.cpp
@@ -1,11 +1,12 @@
 #include <string>
 #include <iostream>
 #include <vector>
+#include <algorithm>
 #include "../headers/formatting.hpp"
 
 void removeSingleAnn(std::vector<std::string> &fileContents, std::string ann) {
-    for (int i = 0; i < fileContents.size(); i++) {
-        if (fileContents[i].find(ann) != -1) {
+    for (size_t i = 0; i < fileContents.size(); i++) {
+        if (fileContents[i].find(ann) != (size_t)-1) {
             fileContents[i] = fileContents[i].substr(0, fileContents[i].find(ann));
         }
     }
@@ -13,14 +14,14 @@ void removeSingleAnn(std::vector<std::string> &fileContents, std::string ann) {
 void removeMultiAnn(std::vector<std::string> &fileContents, std::string ann) {
     size_t pos;
     size_t pos2;
-    for (int i = 0; i < fileContents.size(); i++) {
-        if ((pos=fileContents[i].find(ann)) != -1) {
-            if ((pos2=fileContents[i].find(ann, pos + ann.length())) != -1) {
+    for (size_t i = 0; i < fileContents.size(); i++) {
+        if ((pos=fileContents[i].find(ann)) != (size_t)-1) {
+            if ((pos2=fileContents[i].find(ann, pos + ann.length())) != (size_t)-1) {
                 fileContents[i] = fileContents[i].substr(0, pos) + fileContents[i].substr(pos2 + ann.length());
             } else {
                 fileContents[i] = fileContents[i].substr(0, pos);
                 i++;
-                while ((pos=fileContents[i].find(ann)) == -1) {
+                while ((pos=fileContents[i].find(ann)) == (size_t)-1) {
                     fileContents.erase(fileContents.begin() + i);
                     i++;
                 }
@@ -31,7 +32,7 @@ void removeMultiAnn(std::vector<std::string> &fileContents, std::string ann) {
 }
 
 void removeEmptyLines(std::vector<std::string> &fileContents) {
-    for (int i = 0; i < fileContents.size(); i++) {
+    for (size_t i = 0; i < fileContents.size(); i++) {
 		if (fileContents[i].empty()) {
 			fileContents.erase(fileContents.begin() + i);
 			i--;
@@ -40,15 +41,15 @@ void removeEmptyLines(std::vector<std::string> &fileContents) {
 }
 
 void replaceWinNewLines(std::vector<std::string> &fileContents) {
-    for (int i = 0; i < fileContents.size(); i++) {
+    for (size_t i = 0; i < fileContents.size(); i++) {
 		fileContents[i].erase(std::remove(fileContents[i].begin(), fileContents[i].end(), '\r'), fileContents[i].end());
 	}
 }
 
 void removeBlankLines(std::vector<std::string> &fileContents) {
     bool removeline = true;
-	for (int i = 0; i < fileContents.size(); i++) {
-		for (int j = 0; j < fileContents[i].length(); j++) if (fileContents[i][j] != ' ' && fileContents[i][j] != '\t') removeline = false;
+	for (size_t i = 0; i < fileContents.size(); i++) {
+		for (size_t j = 0; j < fileContents[i].length(); j++) if (fileContents[i][j] != ' ' && fileContents[i][j] != '\t') removeline = false;
 		if (removeline) {
 			fileContents.erase(fileContents.begin() + i);
 			i--;
@@ -60,8 +61,8 @@ void removeBlankLines(std::vector<std::string> &fileContents) {
 std::pair<bool,bool> isSpaced(std::vector<std::string> &fileContents) {
     bool isSpaced = false;
 	bool error = false;
-	for (int i = 0; i < fileContents.size(); i++) {
-		if (fileContents[i].find(":") != -1) {
+	for (size_t i = 0; i < fileContents.size(); i++) {
+		if (fileContents[i].find(":") != (size_t)-1) {
 			if (fileContents[i+1][0] == ' ') isSpaced = true;
             else if (fileContents[i+1][0] != '\t') error = true;
 		}
