+++ src/formatting/languages/python/formatter.cpp
@@ -25,9 +25,9 @@ void pyFormat(std::vector<std::string> &fileContents) {
 		std::vector<int> spacesVec;
 		int indentation = 0;
 		std::cout << "Detected spaced indentation" << std::endl;
-		for (int i = 0; i < fileContents.size(); i++) {
+		for (size_t i = 0; i < fileContents.size(); i++) {
 			spaces = 0;
-			for (int j = 0; j < fileContents[i].length(); j++) {
+			for (size_t j = 0; j < fileContents[i].length(); j++) {
 				if (fileContents[i][j] == ' ') while (fileContents[i][j] == ' ') {
 					spaces++;
 					j++;
@@ -36,7 +36,7 @@ void pyFormat(std::vector<std::string> &fileContents) {
 			}
 			spacesVec.push_back(spaces);
 		}
-		for (int i = 0; i < spacesVec.size(); i++) {
+		for (size_t i = 0; i < spacesVec.size(); i++) {
 			if (spacesVec[i] > 0) {
 				for (int j = stoi(getValue("MAX_INDENT_SPACE")); j > 1; j--) {
 					if (spacesVec[i] % j == 0) {
@@ -56,7 +56,7 @@ void pyFormat(std::vector<std::string> &fileContents) {
 			std::cin >> indentation;
 		}
 		// check manually because i dont trust the user
-		for (int i = 0; i < spacesVec.size(); i++) {
+		for (size_t i = 0; i < spacesVec.size(); i++) {
 			if (spacesVec[i] % indentation != 0) {
 				std::cout << "Indentation error" << std::endl;
 				return;
